feat(145406): Refatorar SalvarLotesException e atualizar o tratamento de erros na função salvar_lotes
