+++ candidatos/service/lotes_service.py
@@ -10,18 +10,9 @@ from django.db import transaction
 
 from candidatos.models import ConcursoCandidato
 from candidatos.models.candidato import Candidato
+from candidatos.service.exceptions import SalvarLotesException
 
 logger = logging.getLogger(__name__)
-
-
-class SalvarLotesException(Exception):
-    def __init__(self, mensagem: str, erros_por_linha: list[dict[str, Any]] | None = None):
-        super().__init__(mensagem)
-        self.mensagem = mensagem
-        self.erros_por_linha = erros_por_linha or []
-
-    def __str__(self) -> str:
-        return self.mensagem
 
 
 @transaction.atomic
@@ -34,25 +25,12 @@ def salvar_lotes(concurso_uuid: str, lotes: list[dict[str, Any]]) -> int:
     3. Atualiza ConcursoCandidato (numero_lote, codigo_sigpec, numero_vaga)
        e Candidato (registro_funcional, vinculo).
     4. Retorna total de candidatos atualizados.
-    """ 
+    """
 
-    erros_por_linha: list[dict[str, Any]] = []
+    erros: list[str] = []
 
-    primeiro_lote = lotes[0]
-    try:
-        nro_lote_reset = int(primeiro_lote.get('lote'))
-    except (TypeError, ValueError):
-        raise SalvarLotesException(
-            'Campo lote invalido na linha 1.',
-            erros_por_linha=[
-                {
-                    'linha': 1,
-                    'identificacao': str(primeiro_lote.get('identificacao', '')).strip(),
-                    'mensagem': 'Campo lote deve ser numerico.',
-                }
-            ],
-        )
-    
+    nro_lote_reset = lotes[0]["lote"]
+
     # 1. Reset: zera campos de lote dos registros com o mesmo numero_lote no concurso
     if nro_lote_reset:
         ConcursoCandidato.objects.filter(
@@ -90,45 +68,28 @@ def salvar_lotes(concurso_uuid: str, lotes: list[dict[str, Any]]) -> int:
                 concurso_uuid,
                 linha,
             )
-            erros_por_linha.append(
-                {
-                    'linha': linha,
-                    'identificacao': identificacao,
-                    'mensagem': f'Candidato com a identificacao {identificacao} nao encontrado.',
-                }
-            )
+            erros.append(f"Linha: {linha} - Candidato com codigo_inscricao: {identificacao} não encontrado")
             continue
 
-        try:
-            cc.numero_lote = int(item.get('lote')) if item.get('lote') is not None else None
-            cc.codigo_sigpec = int(item.get('empresa')) if item.get('empresa') is not None else None
-            cc.numero_vaga = int(item.get('vaga')) if item.get('vaga') is not None else None
-            chave_inscrito = str(item.get('chave_inscrito', '')).strip()
-            cc.chave_inscrito = chave_inscrito or None
-        except (TypeError, ValueError):
-            erros_por_linha.append(
-                {
-                    'linha': linha,
-                    'identificacao': identificacao,
-                    'mensagem': 'Campos lote, empresa e vaga devem ser numericos.',
-                }
-            )
-            continue
+        cc.numero_lote = item["lote"]
+        cc.codigo_sigpec = item["empresa"]
+        cc.numero_vaga = item["vaga"]
+        chave_inscrito = str(item.get('chave_inscrito', '')).strip()
+        cc.chave_inscrito = chave_inscrito or None
 
         cc_para_atualizar.append(cc)
 
-        candidato = cc.candidato
-        if candidato:
-            candidato.registro_funcional = str(item.get('numfunc', '')).strip()
-            candidato.vinculo = str(item.get('numvinc', '')).strip()
-            candidatos_para_atualizar.append(candidato)
+        if cc.candidato:
+            cc.candidato.registro_funcional = item["numfunc"]
+            cc.candidato.vinculo = item["numvinc"]
+            candidatos_para_atualizar.append(cc.candidato)
 
         total_atualizados += 1
 
-    if erros_por_linha:
+    if erros:
         raise SalvarLotesException(
             mensagem='Falha ao salvar lotes. Nenhuma alteracao foi persistida.',
-            erros_por_linha=erros_por_linha,
+            detalhes="\n".join(erros),
         )
 
     if cc_para_atualizar:
